module7-task3: * move db folder
