+++ src/db/models/offer-type/offer-type.ts
