+++ src/db/models/offer/offer.ts
